Feat: 대충 많은 기능

+++ backend/src/main/java/com/ssafy/enjoytrip/features/tripost/adapter/in/web/tripost/response/SearchHotTripostResponse.java
@@ -15,5 +15,5 @@ public class SearchHotTripostResponse {
     private String endWaypoint;
     private String imageUrl;
     private Integer waypointLength;
-    private Integer likeCount;
+    private long likeCount;
 }
