+++ backend/src/main/java/com/ssafy/enjoytrip/features/tripost/adapter/in/web/tripost/response/SearchLatestTripostResponse.java
@@ -15,6 +15,6 @@ public class SearchLatestTripostResponse {
     private String description;
     private LocalDate createdAt;
     private String imageUrl;
-    private Integer viewCount;
-    private Integer commentCount;
+    private long viewCount;
+    private long commentCount;
 }
